+++ CHANGELOG.md
@@ -7,3 +7,4 @@
   * Auto-install CLI plugins for running services during `dcos cluster setup` when environment variable `DCOS_CLI_EXPERIMENTAL_AUTOINSTALL_PACKAGE_CLIS` set
   * New command `dcos config keys` printing all the keys that can be set in a configuration file
   * The new `dcos cluster open` can be used to open the currently attached cluster UI in the browser.
+  * Log into OSS without copy/pasting authentication token from browser.